Refactor : 사용자(회원정보수정) - 패스워드 단일 DTO로 변경 (Issue #3)
UpdateUserRequestDto(nickname, profile_url, password) -> UpdatePwdRequestDto(password)
